chore: packaging, CLI entry points, and validation cleanup
1. Move non-core dependencies to optional groups — only sqlglot is
   required at runtime.  pytest goes to [dev], dotenv/langfuse/
   matplotlib/networkx go to [scripts].

2. Move CLI into the package (src/fastssv/cli.py) and add
   [project.scripts] + __main__.py so users can run the validator
   via `fastssv` or `python -m fastssv` after install.  Root main.py
   is kept as a thin backwards-compat wrapper.

3. Deduplicate the 6 validate_* convenience functions in
   rules/__init__.py into a shared _validate_category() helper.
   SQL is still parsed once up-front for error reporting, but the
   copy-pasted boilerplate is gone.

4. Add PEP 561 py.typed marker with package-data config so downstream
   type checkers pick up the inline type hints.
