+++ pyproject.toml
@@ -6,13 +6,22 @@ readme = "README.md"
 license = {text = "Apache-2.0"}
 requires-python = ">=3.12"
 dependencies = [
+	"sqlglot>=24.6.0",
+]
+
+[project.optional-dependencies]
+dev = [
+	"pytest>=9.0.3",
+]
+scripts = [
 	"dotenv>=0.9.9",
 	"langfuse>=3.12.1",
 	"matplotlib>=3.10.8",
 	"networkx>=3.6.1",
-	"pytest>=9.0.3",
-	"sqlglot>=24.6.0",
 ]
+
+[project.scripts]
+fastssv = "fastssv.cli:main"
 
 [build-system]
 requires = ["setuptools>=68", "wheel"]
@@ -20,6 +29,9 @@ build-backend = "setuptools.build_meta"
 
 [tool.setuptools]
 package-dir = {"" = "src"}
+
+[tool.setuptools.package-data]
+fastssv = ["py.typed"]
 
 [tool.setuptools.packages.find]
 where = ["src"]
