+++ src/fastssv/__main__.py
@@ -1,0 +1,5 @@
+"""Allow running fastSSV as ``python -m fastssv``."""
+
+from fastssv.cli import main
+
+raise SystemExit(main())
